# fastingmate-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Dokumentasi API untuk autentikasi pengguna",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/food-logs": {
            "get": {
                "tags": [
                    "Food Logs"
                ],
                "summary": "Menampilkan daftar konsumsi makanan pengguna",
                "operationId": "7bdc2857a1cf61dcaa1bc399b4483727",
                "responses": {
                    "200": {
                        "description": "Daftar konsumsi makanan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "food_id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "quantity": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1.5
                                            },
                                            "meal_type": {
                                                "type": "string",
                                                "example": "lunch"
                                            },
                                            "total_calories": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 500
                                            },
                                            "total_protein": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 30
                                            },
                                            "total_carbs": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 60
                                            },
                                            "total_fat": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 10
                                            },
                                            "consumed_at": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2025-03-21 12:00:00"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2025-03-21 12:05:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Food Logs"
                ],
                "summary": "Menambahkan catatan konsumsi makanan baru",
                "operationId": "cbb78675c0147f34f2ae31d1f3e31c41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "food_id",
                                    "quantity",
                                    "meal_type",
                                    "total_calories",
                                    "total_protein",
                                    "total_carbs",
                                    "total_fat",
                                    "consumed_at"
                                ],
                                "properties": {
                                    "food_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.5
                                    },
                                    "meal_type": {
                                        "type": "string",
                                        "example": "lunch"
                                    },
                                    "total_calories": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "total_protein": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30
                                    },
                                    "total_carbs": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 60
                                    },
                                    "total_fat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "consumed_at": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-03-21 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catatan konsumsi makanan berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/food-logs/{id}": {
            "delete": {
                "tags": [
                    "Food Logs"
                ],
                "summary": "Menghapus catatan konsumsi makanan berdasarkan ID",
                "operationId": "0efb29c9f75c31a7cd868232c92b65d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catatan konsumsi makanan berhasil dihapus"
                    },
                    "404": {
                        "description": "Catatan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <your-token>)"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}